requirements.txt
openai==1.12.0
python-dotenv==1.0.0
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
Pillow==10.2.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9 